+++ src/main/java/com/javerian/erp/mlm/service/auth/UserService.java
@@ -6,7 +6,7 @@ import com.javerian.erp.mlm.model.auth.User;
 
 public interface UserService {
 
-	User findById(int id);
+	User findById(Long id);
 
 	User findBySSO(String sso);
 
@@ -18,6 +18,6 @@ public interface UserService {
 
 	List<User> findAllUsers();
 
-	boolean isUserSSOUnique(Integer id, String sso);
+	boolean isUserSSOUnique(Long id, String sso);
 
 }
